Update README.md
Q1. Open the data LotteryExample.dta that I posted on Canvas (note: LotteryExample_s13.dta can be opened in Stata 13 and is posted under the [data] folder). For this dataset, I created 100,000 fake student records that include the following variables: parent_income, ability, lottery, attends, and test_score. They are defined as follows:

parent_income: total parent income earned in the year before charter school admission.

ability: a measure of the student’s intellectual abilities.

lottery: equal to one if the student won the admissions lottery, and it equals zero if not.

attends: equal to one if the student attends the charter school, and it equals zero if not.

test_score: student’s standardized test score at the end of 12th grade.